Use sha of validated commit to publish snapshot
Later jobs also know this sha (it's propagated by jenkins),
so that they can easily resolve it. Currently used by IDE validation.

Before, we were using the sha of current HEAD, which is the merge
commit of merging the tested commit into the target branch.
This is not information readily available in downstream jobs, however.