Speed up PR validation by publishing locker to maven from the start.
Use the refactorings in https://github.com/scala/scala/issues/3108 and
https://github.com/scala/scala/issues/3109 to speed up builds.

`pr-scala-publish-core` does a quick build, skipping locker
(effectively using starr for locker), and publishes that result.

`pr-scala-test` then builds quick using starr (skipping locker), where
the starr version is the compiler just published (which corresponds
to locker in the usual layering, so we actually get the usual layering).

Standardize on `versions.properties`, which is produced by `ant` as
`scala/buildcharacter.properties`, and moved to `/versions.properties`
by `pr-scala-publish-core`. This has all the info to characterize
the Scala build and all the modules that are part of it.

Also, one small refactoring: make pr-scala-test reusable for rangepos