Pass updated scalacheck.version.number and jline.version to the core build
Just like AKKA_ACTOR_VER, JLINE_VER can be provided to use a new jline
release.

For all modules, the upgraded version number is also forwareded to
the core build if the new module version already existed. This way
we can upgrade versions.properties to a new module release that
is already published.